Extract apps plugins from core to separate packages
