+++ app/background/rpc/initialize.js
@@ -1,7 +1,5 @@
 import register from 'lib/rpc/register'
 
-import getFileDetails from 'lib/getFileDetails'
-import getFileSize from 'lib/getFileSize'
 import initializePlugins from 'lib/initializePlugins'
 
 /**
@@ -9,7 +7,5 @@ import initializePlugins from 'lib/initializePlugins'
  * After `register` function can be called using rpc from main window
  */
 export default () => {
-  register('getFileDetails', getFileDetails)
-  register('getFileSize', getFileSize)
   register('initializePlugins', initializePlugins)
 }
