--- app/main/components/FileDetails/styles.css
@@ -1,31 +1,0 @@
-.fileName {
-  font-size: 1.5em;
-  font-weight: normal;
-  text-align: center;
-  max-width: 100%;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-.fileDetails {
-  margin-top: 50px;
-  font-size: .8em;
-  color: var(--secondary-font-color);
-  width: 100%;
-
-  dt {
-    float: left;
-    width: 49%;
-    padding: 0;
-    margin: 0;
-    text-align: right;
-    margin-right: 2%;
-  }
-
-  dd {
-    float: left;
-    width: 49%;
-    padding: 0;
-    margin: 0;
-    text-align: left;
-  }
-}
