--- app/main/plugins/core/basic-apps/styles.css
@@ -1,13 +1,0 @@
-.previewIcon {
-  text-align: center;
-
-  img {
-    max-width: 200px;
-    max-height: 200px;
-  }
-}
-
-.previewName {
-  text-align: center;
-  font-size: 1.2em;
-}
