--- app/main/plugins/core/cerebro/index.js
@@ -1,10 +1,0 @@
-import autocomplete from './autocomplete'
-import quit from './quit'
-import plugins from './plugins'
-import settings from './settings'
-import version from './version'
-import reload from './reload'
-
-export {
-  autocomplete, quit, plugins, settings, version, reload
-}
