+++ app/main/plugins/core/icon.png
