--- app/main/plugins/core/osx-apps/styles.css
@@ -1,19 +1,0 @@
-:global {
-  .gm-style {
-    filter: var(--preview-filter);
-  }
-}
-
-.previewIcon {
-  text-align: center;
-
-  img {
-    max-width: 200px;
-    max-height: 200px;
-  }
-}
-
-.previewName {
-  text-align: center;
-  font-size: 1.2em;
-}
