+++ app/main/plugins/core/plugins/Preview/ActionButton.js
