+++ app/main/plugins/core/plugins/Preview/Settings.js
