+++ app/main/plugins/core/plugins/Preview/index.js
