+++ app/main/plugins/core/plugins/Preview/styles.css
