+++ app/main/plugins/core/plugins/format.js
