+++ app/main/plugins/core/plugins/getAvailablePlugins.js
