+++ app/main/plugins/core/plugins/getInstalledPlugins.js
