+++ app/main/plugins/core/plugins/getReadme.js
