+++ app/main/plugins/core/plugins/index.js
