+++ app/main/plugins/core/plugins/initializeAsync.js
