+++ app/main/plugins/core/plugins/loadPlugins.js
