+++ app/main/plugins/core/quit/index.js
