+++ app/main/plugins/core/reload/index.js
