+++ app/main/plugins/core/settings/Settings/Hotkey.js
