+++ app/main/plugins/core/settings/Settings/index.js
