+++ app/main/plugins/core/settings/Settings/styles.css
