+++ app/main/plugins/core/settings/index.js
