+++ app/main/plugins/core/version/index.js
