+++ package.json
@@ -198,9 +198,7 @@
     "cerebro-tools": "0.1.8",
     "cerebro-ui": "0.0.13",
     "co": "4.6.0",
-    "du": "0.1.0",
     "escape-string-regexp": "1.0.5",
-    "event-stream": "3.3.4",
     "fix-path": "2.1.0",
     "glob": "7.0.5",
     "lodash": "4.13.1",
